+++ ui/packages/components/src/Button/buttonStyles.ts
@@ -1,13 +1,10 @@
-import type { ButtonAppearance, ButtonKind, ButtonSize } from './Button';
-
 interface ButtonColorParams {
-  kind: ButtonKind;
-  appearance: ButtonAppearance;
-  loading?: boolean;
+  kind: 'default' | 'primary' | 'success' | 'danger';
+  appearance: 'solid' | 'outlined' | 'text';
 }
 
 interface ButtonSizeParams {
-  size: ButtonSize;
+  size: 'small' | 'regular' | 'large';
 }
 
 interface ButtonSizeStyleParams extends ButtonSizeParams {
@@ -15,39 +12,38 @@ interface ButtonSizeStyleParams extends ButtonSizeParams {
   label?: React.ReactNode;
 }
 
-export const getButtonColors = ({ kind, appearance, loading }: ButtonColorParams) => {
+export const getButtonColors = ({ kind, appearance }: ButtonColorParams) => {
   const solidButtonStyles = {
-    primary: loading
-      ? 'bg-primary-xSubtle text-onContrast'
-      : 'bg-primary-intense focus:bg-primary-xIntense hover:bg-primary-xIntense active:bg-primary-2xIntense disabled:bg-primary-xSubtle text-onContrast',
-    secondary: '', // NOOP: there are no designs for secondary solid buttons,
-    danger: loading
-      ? 'bg-tertiary-xSubtle text-onContrast'
-      : 'bg-tertiary-intense focus:bg-primary-xIntense hover:bg-tertiary-xIntense active:bg-tertiary-2xIntense disabled:bg-tertiary-xSubtle text-onContrast',
+    default:
+      'bg-slate-800 border-t border-white/10 hover:bg-slate-800/80 text-slate-100 hover:text-white',
+    primary:
+      'bg-indigo-500 border-t border-white/10 hover:bg-indigo-500/80 text-slate-100 hover:text-white',
+    success:
+      'bg-emerald-600 border-t border-white/10 hover:bg-emerald-600/80 text-slate-100 hover:text-white',
+    danger:
+      'bg-rose-700 border-t border-white/10 hover:bg-rose-700/80 text-slate-100 hover:text-white',
   };
 
   const outlinedButtonStyles = {
-    primary: loading
-      ? 'border border-subtle text-primary-moderate'
-      : 'border border-muted text-primary-intense focus:bg-subtle hover:bg-subtle active:bg-muted disabled:bg-disabled disabled:text-primary-xSubtle',
-    secondary: loading
-      ? 'border border-subtle text-foreground-subtle'
-      : 'border border-muted text-foreground-base focus:bg-subtle hover:bg-subtle active:bg-muted disabled:bg-disabled disabled:text-foreground-disabled',
-    danger: loading
-      ? 'border border-subtle text-tertiary-moderate'
-      : 'border border-muted text-tertiary-intense focus:bg-subtle hover:bg-subtle active:bg-muted disabled:bg-disabled disabled:text-tertiary-xSubtle',
+    default:
+      'bg-white dark:bg-slate-800/20 border dark:border-slate-800/80 hover:dark:border-slate-800 text-slate-700 dark:text-slate-200 hover:bg-slate-100 hover:text-indigo-500 dark:hover:text-white',
+    primary:
+      'bg-white dark:bg-indigo-500/20 border border-indigo-500/80 hover:border-indigo-500 text-indigo-500 dark:text-slate-200 hover:bg-slate-100 dark:hover:text-white',
+    success:
+      'bg-white dark:bg-emerald-600/20 border border-emerald-600/80 hover:border-emerald-600 text-emerald-600 dark:text-slate-200 hover:bg-slate-100 dark:hover:text-white',
+    danger:
+      'bg-white dark:bg-rose-700/20 border border-rose-700/80 hover:border-rose-700 text-rose-700 dark:text-slate-200 hover:bg-slate-100 dark:hover:text-white',
   };
 
-  const ghostButtonStyles = {
-    primary: loading
-      ? 'text-primary-moderate'
-      : 'text-primary-intense focus:bg-subtle hover:bg-subtle active:bg-muted disabled:text-primary-xSubtle',
-    secondary: loading
-      ? 'text-foreground-subtle'
-      : 'text-foreground-base focus:bg-subtle hover:bg-subtle active:bg-muted disabled:text-foreground-disabled',
-    danger: loading
-      ? 'text-tertiary-moderate'
-      : 'text-tertiary-intense focus:bg-subtle hover:bg-subtle active:bg-muted disabled:bg-disabled disabled:text-tertiary-xSubtle',
+  const textButtonStyles = {
+    default:
+      'text-slate-500 hover:text-slate-500/80 dark:text-slate-400 dark:hover:text-slate-400/80 hover:underline',
+    primary:
+      'text-indigo-500 hover:text-indigo-500/80 dark:text-indigo-400 hover:text-indigo-400/80 hover:underline',
+    success:
+      'text-emerald-600 hover:text-emerald-600/80 dark:text-emerald-500 dark:hover:text-emerald-500/80 hover:underline',
+    danger:
+      'text-rose-500 hover:text-rose-500/80 dark:text-rose-400 dark:hover:text-rose-400/80 hover:underline',
   };
 
   if (appearance === 'solid') {
@@ -55,17 +51,18 @@ export const getButtonColors = ({ kind, appearance, loading }: ButtonColorParams
   } else if (appearance === 'outlined') {
     return outlinedButtonStyles[kind];
   } else {
-    return ghostButtonStyles[kind];
+    return textButtonStyles[kind];
   }
 };
 
 export const getKeyColor = ({ appearance, kind }: ButtonColorParams) => {
   const defaultKeyStyles = {
+    default: 'bg-slate-800/80',
     primary: 'bg-indigo-500/80',
-    secondary: 'bg-emerald-600/80',
+    success: 'bg-emerald-600/80',
     danger: 'bg-rose-700/80',
   };
-  if (appearance === 'solid' && kind === 'primary') {
+  if (appearance === 'solid' && kind === 'default') {
     return 'bg-slate-900';
   } else if (appearance === 'solid') {
     return 'bg-slate-800/20';
@@ -77,21 +74,21 @@ export const getKeyColor = ({ appearance, kind }: ButtonColorParams) => {
 
 export const getButtonSizeStyles = ({ size, icon, label }: ButtonSizeStyleParams) => {
   const iconOnlySizeStyles = {
-    small: 'h-6 p-1.5',
-    medium: 'h-8 p-1.5',
-    large: 'h-10 p-1.5',
+    small: 'w-7 h-7',
+    regular: 'w-8 h-8',
+    large: 'w-10 h-10',
   };
 
   const sizeStyles = {
-    small: 'h-6 text-xs leading-[18px] px-3 py-1.5',
-    medium: 'h-8 text-xs leading-[18px] px-3 py-1.5',
-    large: 'h-10 text-xs leading-[18px] px-3 py-1.5',
+    small: 'text-xs px-2.5 h-7 leading-7',
+    regular: 'text-sm px-2.5 h-8 leading-8',
+    large: 'text-base px-2.5 h-10 leading-10',
   };
 
   return icon && !label ? iconOnlySizeStyles[size] : sizeStyles[size];
 };
 
-export const getDisabledStyles = ({ appearance }: ButtonColorParams) => {
+export const getDisabledStyles = ({ appearance, kind }: ButtonColorParams) => {
   if (appearance === 'solid') {
     return 'disabled:cursor-not-allowed disabled:text-slate-400 disabled:bg-slate-200 dark:disabled:text-slate-500 dark:disabled:bg-slate-800 ';
   } else if (appearance === 'outlined') {
@@ -102,9 +99,9 @@ export const getDisabledStyles = ({ appearance }: ButtonColorParams) => {
 
 export const getIconSizeStyles = ({ size }: ButtonSizeParams) => {
   const sizeStyles = {
-    small: 'h-4 w-4',
-    medium: 'h-4 w-4',
-    large: 'h-4 w-4',
+    small: 'h-3.5 w-3.5',
+    regular: 'h-4 w-4',
+    large: 'h-5 w-5',
   };
 
   return sizeStyles[size];
@@ -112,12 +109,13 @@ export const getIconSizeStyles = ({ size }: ButtonSizeParams) => {
 
 export const getSpinnerStyles = ({ appearance, kind }: ButtonColorParams) => {
   const defaultSpinnerStyles = {
-    primary: 'fill-primary-moderate',
-    secondary: 'fill-foreground-subtle',
-    danger: 'fill-tertiary-moderate',
+    default: 'fill-slate-700 dark:fill-white',
+    primary: 'fill-indigo-500 dark:fill-white',
+    success: 'fill-emerald-600 dark:fill-white',
+    danger: 'fill-rose-700 dark:fill-white',
   };
   if (appearance === 'outlined') {
     return defaultSpinnerStyles[kind];
   }
-  return 'fill-foreground-onContrast';
+  return 'fill-white';
 };
