+++ ui/packages/components/src/Button/index.ts
@@ -1,4 +1,3 @@
-export { OldButton as Button } from './OldButton';
-export { Button as NewButton } from './Button';
-export { getButtonColors, getButtonSizeStyles, getDisabledStyles } from './oldButtonStyles';
+export { Button } from './Button';
+export { getButtonColors, getButtonSizeStyles, getDisabledStyles } from './buttonStyles';
 export { SplitButton } from './SplitButton';
